+++ lib/models/source.dart
@@ -1,4 +1,4 @@
-import 'package:hive/hive.dart';
+import 'package:hive_ce/hive.dart';
 import 'track.dart';
 
 part 'source.g.dart';