+++ test/widgets/show_list_item_details_test.dart
@@ -12,7 +12,7 @@ import 'package:shakedown/providers/show_list_provider.dart';
 import 'package:shakedown/providers/theme_provider.dart';
 import 'package:shakedown/services/catalog_service.dart';
 import 'package:shakedown/models/rating.dart';
-import 'package:hive_flutter/hive_flutter.dart';
+import 'package:hive_ce_flutter/hive_flutter.dart';
 import 'package:flutter/foundation.dart';
 import 'package:mockito/mockito.dart';
 import 'package:shakedown/services/device_service.dart';